spells: widen spell-ID types from uint16_t to uint32_t
spells_us.txt now goes to 73999 and Live spell IDs exceed 65535.
The wire (charProfileStruct.buffs[].spellid is int32_t, OP_Buff
spellID is uint32_t) and the proto (spell_id is uint32) already
carry 32-bit IDs; only the SpellShell/Spells/util layer was 16-bit,
truncating any future ID >= 0x10000.

Widened:
  SpellItem::{spellId,setSpellId,update,m_spellId}
  SpellShell::findSpell (both overloads)
  Spell::{spell,m_spell}
  Spells::{spell,maxSpell,m_maxSpell}
  spell_name(util)
  Player::m_lastSpellOnId (declared, unused — kept consistent)
  combatrouter cast from int32 wire field to uint32

+++ src/combatrouter.cpp
@@ -41,7 +41,7 @@ void CombatRouter::action2(const uint8_t* data, size_t len, uint8_t /*dir*/)
 
     QString spellName;
     if (a->spell > 0 && m_spells) {
-        if (const Spell* sp = m_spells->spell(static_cast<uint16_t>(a->spell))) {
+        if (const Spell* sp = m_spells->spell(static_cast<uint32_t>(a->spell))) {
             spellName = sp->name();
         }
     }
